Fix compilation warnings (#12)
* No warnings on mac!

* Still fixing clang-tidy errors. Lots of comparisons between longs and ints that will cause problems.

* Half way done, modified a lot of deduced type conversions. May need to go back in the future to remove them.

* Should fix all the warnings on linux and windows.

* Pushing through more deduced type fixes.

* Should be the final fix for linux.

+++ src/ClusterExportTestFixture.cpp
@@ -9,7 +9,7 @@ bool ClusterExportTestFixture::TestSetListVector(ListVector& listVector, const s
     clusteExport->SetListVector(listVector, label);
     const auto pair = clusteExport->GetListVector();
     const std::string otherLabel = pair.label;
-    const size_t otherSize = pair.listVector->size();
+    const int otherSize = pair.listVector->size();
     const bool finalResult = result == (otherLabel == label && otherSize == listVector.size());
     TearDown();
     return finalResult;
@@ -20,7 +20,7 @@ bool ClusterExportTestFixture::TestGetListVector(const bool result) {
     auto* vector = new ListVector();
     clusteExport->SetListVector(*vector, "0.00");
     const auto pair = clusteExport->GetListVector();
-    const size_t otherSize = pair.listVector->size();
+    const int otherSize = pair.listVector->size();
     const bool finalResult = result == (otherSize == vector->size());
     TearDown();
     return finalResult;
